pyproject.toml
[project]
name = "interpr-credit-risk"
version = "0.1.0"
description = "Reproducible SHAP vs LIME study for credit-risk default prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.16.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "pyyaml>=6.0.2",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "scikit-learn>=1.5.1",
  "imbalanced-learn>=0.12.3",
  "xgboost>=2.1.1",
  "lightgbm>=4.5.0",
  "catboost>=1.2.7",
  "shap>=0.46.0",
  "lime>=0.2.0.1",
  "matplotlib>=3.9.2",
  "seaborn>=0.13.2",
  "scipy>=1.14.1",
  "joblib>=1.4.2",
  "rich>=13.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "ruff>=0.6.2",
]

[project.scripts]
icr = "icr.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
